migdalproject/migdal_platform/requirements.txt
# --- Core Web Framework ---
Django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-environ==0.12.0
psycopg2-binary==2.9.11

# --- Security & Licensing ---
PyJWT==2.10.1
cryptography==46.0.4
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
asn1crypto==1.5.1
oscrypto==1.3.0

# --- Google AI / API Integration ---
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.61.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0

# --- PDF & Document Generation ---
reportlab==4.4.9
xhtml2pdf==0.2.17
pypdf==6.6.2
svglib==1.6.0
pycairo==1.29.0
rlPyCairo==0.4.0

# --- Data Processing ---
pandas==3.0.0
numpy==2.4.1
lxml==6.0.2

# --- Network & Async ---
requests==2.32.5
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
websockets==15.0.1
urllib3==2.6.3

# --- Utilities & Misc ---
pillow==12.1.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
sqlparse==0.5.5
typing_extensions==4.15.0
pydantic==2.12.5
pydantic_core==2.41.5
tzdata==2025.3
tzlocal==5.3.1
colorama==0.4.6
tqdm==4.67.1

# --- Production Server ---
gunicorn==25.1.0
